sync: add `OnceCell::wait_initialized` method
Such a method was mentioned in #7151 as nice to have, so here it is.
The previous `OnceCell` implementation was using a semaphore, but if it
was convenient to synchronize initialization, it is not suited to wait
for it. That's why the synchronization has completely been rewritten to
handle it.

There is now 3 possible states: *unset*, *locked*, and *set*. The cell
is initialized to *unset*. When it is initialized, the cell is first
locked, i.e. its state set to *locked*, then the value is written, and
finally the state is set to *set*. As the other synchronization
primitives, it uses a mutex-protected linked list of waiters, which is
drained when initialization succeeds. If initialization fails, the
state is rolled back to *unset* and one awaiting initializer is wakened
up.

The new implementation takes one less word in memory, and it has been
optimized for ZST case. In fact, another goal of this new
implementation is to serve as base for the `Event` primitive of #7151,
so basically a `OnceCell<()>`.
By the way, a proper tracing implementation has been added to
`OnceCell`.