time: rename `cached_when` to `registered_when`
The word `cache` usually hints loading something
is expensive, so the `cached_when` is likely
making developer think this field cached the
`StateCell::state` as loading it is expensive.

But actually,

* The `cached_when` is only for locating the `TimerEntry`
  in the intrusive linked list.
* The `StateCell::state` is the single source of truth
  while picking the expired `TimerEntry`s.
* Loading the `StateCell:state` is also an relaxed load,
  which is the same operation as loading the `cached_when`,
  and it not more expensive than loading `cached_when`.

So the better name of `cached_when` is `registered_when`,
which explicitly tells developers that this field records
the time when the `TimerEntry` was registered.