Introduce the `SetOnce` struct
The `SetOnce` data structure allows you to contain a T value and wait
until the initialization of that T value in the Cell. It can only be
initialized once and can be destroyed using `into_inner` or by dropping
it.

This is equivalent to the `asyncioEvent` class https://docs.python.org/3/library/asyncio-sync.html#asyncio.Event`

The wait() allows you to block on a future which wakes up when the
`SetOnce` is initalized.

Added tokio/tests/sync_set_once.rs