runtime: Make multithreaded rt LIFO slot atomic
This commit changes the LIFO slot on multi-threaded runtime workers from
being a mutable `Option<Notified<Arc<Handle>>>` to a new
`AtomicNotified` type. This
type implements a cell containing a nullable task pointer which can
be swapped atomically. It's analogous to `AtomicCell` but with the extra
`PhantomData` to remember the task's scheduler type parameter, which
would otherwise be erased by the conversion into a `*mut Header``
pointer.

This change is in preparation for a subsequent change to allow
work-stealing from the LIFO slot (see: #4941).