runtime: add test for #4941
This commit adds a test ensuring that if a task is notified to the LIFO
slot by another task which then blocks the worker thread forever, the
LIFO task is eventually stolen by another worker. I've confimed that
this test fails on the `master` branch, and passes after these changes.