runtime: fix racy `worker_local_queue_depth` test
This test spawns a task that sometimes ends up in the LIFO slot and
sometimes doesn't. This was previously fine as the LIFO slot didn't
count for `worker_local_queue_depth`, but now it does. Thus, we have to
make sure that task no longer exists before asserting about queue depth.